Add training script for new models

+++ dl_car_control/Ackermann/utils/pilot_net_dataset.py
@@ -1,9 +1,9 @@
 from torch.utils.data.dataset import Dataset
 from torchvision import transforms
 from PIL import Image
-from utils.processing import *
+from dl_car_control.Ackermann.utils.processing import *
 from pathlib import Path
-#import inspect
+# import inspect
 
 
 class PilotNetDataset(Dataset):
@@ -18,7 +18,7 @@ class PilotNetDataset(Dataset):
                 type_image = None
             else:
                 type_image = 'cropped'
-            
+
             if 'extreme' in preprocessing:
                 data_type = 'extreme'
             else:
@@ -27,33 +27,33 @@ class PilotNetDataset(Dataset):
             type_image = 'cropped'
             data_type = None
         #print(inspect.getsource(load_data))
-        print('*'*8, "Loading Datasets", '*'*8)    
-        for path in path_to_data: 
+        print('*'*8, "Loading Datasets", '*'*8)
+        for path in path_to_data:
             all_images, all_data = load_data(path)
-            self.images = get_images(all_images, type_image, self.images)        
+            self.images = get_images(all_images, type_image, self.images)
             self.labels = parse_csv(all_data, self.labels)
-        
- 
+
+
         self.labels, self.images = preprocess_data(self.labels, self.images, flip_images, data_type)
 
 
         self.transforms = transforms
-        
+
         self.image_shape = self.images[0].shape
         self.num_labels = np.array(self.labels[0]).shape[0]
 
         self.count = len(self.images)
-        
+
     def __getitem__(self, index):
 
         img = self.images[index]
-        label = np.array(self.labels[index])
+        label = np.array(self.labels[index], dtype=np.float32)
         data = Image.fromarray(img)
-        
+
         if self.transforms is not None:
             data = self.transforms(data)
-        
-        
+
+
         return (data, label)
 
     def __len__(self):
